Just moved some stuff around because it felt more logical
