+++ esque/resources/resource.py
